Move config classes to config package (from util).
(cherry picked from commit ce2306e)
